pytest/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
Pillow==7.2.0
requests==2.23.0
soupsieve==2.0.1
urllib3==1.25.9
